fix: resolve all AUDIT.md findings (pass 1)
Critical bugs fixed:
- fragmented.rs: build_vpcc_fmp4 malformed vpcC box (missing FullBox header,
  wrong bit packing, missing codecInitializationDataSize)
- fragmented.rs: build_hvcc_fmp4 all-zero profile/level → now reads real
  values from HevcConfig (general_profile_idc, general_level_idc, etc.)
- fragmented.rs: build_av1c_fmp4 all-zero codec config → now parses from
  sequence header OBU via extract_av1_config
- muxer/mp4.rs: build_audio_specific_config AudioObjectType always 2 (LC)
  regardless of AacProfile → now correctly sets AOT 1/2/3/4/5/29
- muxer/mp4.rs: build_mdhd_box duration silent u64→u32 truncation for
  durations >13.6h → now uses version=1 (64-bit) path when needed
- codec/h265.rs: is_hevc_keyframe panics on empty input via assert_invariant
  → replaced with graceful early return false

Spec non-compliance fixed:
- muxer/mp4.rs: build_tkhd_box flags=0 disables tracks → now 0x0000_0003
  (track_enabled | track_in_movie) per ISO 14496-12
- muxer/mp4.rs: build_vmhd_box flags=0 → now 0x0000_0001 per ISO 14496-12 §12.1.2
- muxer/mp4.rs: build_mvhd_payload next_track_ID hardcoded 2 even with
  audio → now parameterized (2 for video-only, 3 for video+audio)

Dead invariants removed:
- codec/av1.rs: INV-201 (obu_type & 0x0f can never exceed 15)
- codec/av1.rs: INV-203 (payload non-empty check after prior early return)
- codec/h264.rs: INV-302 (SPS/PPS already filtered by iterator)
- codec/vp9.rs: INV-401 (frame_marker already checked before this point)

API / code quality:
- api.rs: Added Debug impls for Muxer<W> and MuxerBuilder<W>
- api.rs: Fixed MuxerError::FirstVideoFrameMissingSpsPps Display to mention
  H.264 NAL 7/8 and H.265 NAL 32/33/34
- api.rs: Updated write_video doc to cover all supported codecs
- validation.rs: channels field promoted from u8 → u16 to match API
- lib.rs: invariant_ppt marked #[doc(hidden)] (internal module)
- fragmented.rs: encode_language_code deduplicated (canonical in mp4.rs)
- fragmented.rs: FragmentedError marked #[non_exhaustive]
- codec/vp9.rs: Vp9Config now derives Eq
- codec/h265.rs: BLA NAL type doc strings now distinct

Documentation:
- codec/h264.rs: default_avc_config() doc corrected (640x480 Baseline/3.0)
- codec/h265.rs: BLA_W_LP / BLA_W_RADL / BLA_N_LP docs now distinct
- ROADMAP.md: version updated to v0.2.5, test count corrected to 123

Test updates:
- fixtures/minimal.mp4: regenerated after tkhd/vmhd flag corrections
- tests/property_tests.rs: contract tests updated to reflect removed dead
  invariants (INV-201, INV-302, INV-401, INV-503)

+++ src/validation.rs
@@ -70,7 +70,7 @@ pub struct VideoValidationConfig {
 pub struct AudioValidationConfig {
     pub codec: Option<AudioCodec>,
     pub sample_rate: Option<u32>,
-    pub channels: Option<u8>,
+    pub channels: Option<u16>,
     pub sample_frame: Option<Vec<u8>>,
 }
 
@@ -129,7 +129,7 @@ pub fn validate_video_config(
 pub fn validate_audio_config(
     codec: AudioCodec,
     sample_rate: u32,
-    channels: u8,
+    channels: u16,
 ) -> ValidationResult {
     let mut result = ValidationResult::valid();
